% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launches the Sidora shiny app}
\usage{
launch_app()
}
\value{
shiny application object
}
\description{
Launches the Sidora shiny app
}
\examples{
\dontrun{
launchApp()
}

}
